+++ crates/objc2-encode/CHANGELOG.md
@@ -7,7 +7,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ## Unreleased - YYYY-MM-DD
 
 
-## 2.0.0-pre.4 - 2023-02-03
+## 2.0.0-pre.4 - 2023-02-07
 
 ### Added
 * Made the crate `no_std` compatible.